extract test from ConnectivityTest. removed random test
